// src/models/planProyecto.model.ts
export class PlanProyectoModel {
  Id_PlanProyecto: any;
  Codigo_PlanProyecto: String;
  Nombre_PlanProyecto: String;
  EneroSolesP: Number;
  FebreroSolesP: Number;
  MarzoSolesP: Number;
  AbrilSolesP: Number;
  MayoSolesP: Number;
  JunioSolesP: Number;
  JulioSolesP: Number;
  AgostoSolesP: Number;
  SeptiembreSolesP: Number;
  OctubreSolesP: Number;
  NoviembreSolesP: Number;
  DiciembreSolesP: Number;
  EneroDolaresP: Number;
  FebreroDolaresP: Number;
  MarzoDolaresP: Number;
  AbrilDolaresP: Number;
  MayoDolaresP: Number;
  JunioDolaresP: Number;
  JulioDolaresP: Number;
  AgostoDolaresP: Number;
  SeptiembreDolaresP: Number;
  OctubreDolaresP: Number;
  NoviembreDolaresP: Number;
  DiciembreDolaresP: Number;
  EneroSolesG: Number;
  FebreroSolesG: Number;
  MarzoSolesG: Number;
  AbrilSolesG: Number;
  MayoSolesG: Number;
  JunioSolesG: Number;
  JulioSolesG: Number;
  AgostoSolesG: Number;
  SeptiembreSolesG: Number;
  OctubreSolesG: Number;
  NoviembreSolesG: Number;
  DiciembreSolesG: Number;
  EneroDolaresG: Number;
  FebreroDolaresG: Number;
  MarzoDolaresG: Number;
  AbrilDolaresG: Number;
  MayoDolaresG: Number;
  JunioDolaresG: Number;
  JulioDolaresG: Number;
  AgostoDolaresG: Number;
  SeptiembreDolaresG: Number;
  OctubreDolaresG: Number;
  NoviembreDolaresG: Number;
  DiciembreDolaresG: Number;
  Codigo_Proyecto: String;
  Ano: Number;
  Movimiento: Boolean;

  constructor() {
    this.Id_PlanProyecto = null;
    this.Codigo_PlanProyecto = '';
    this.Nombre_PlanProyecto = '';
    this.EneroSolesP = 0;
    this.FebreroSolesP = 0;
    this.MarzoSolesP = 0;
    this.AbrilSolesP = 0;
    this.MayoSolesP = 0;
    this.JunioSolesP = 0;
    this.JulioSolesP = 0;
    this.AgostoSolesP = 0;
    this.SeptiembreSolesP = 0;
    this.OctubreSolesP = 0;
    this.NoviembreSolesP = 0;
    this.DiciembreSolesP = 0;
    this.EneroDolaresP = 0;
    this.FebreroDolaresP = 0;
    this.MarzoDolaresP = 0;
    this.AbrilDolaresP = 0;
    this.MayoDolaresP = 0;
    this.JunioDolaresP = 0;
    this.JulioDolaresP = 0;
    this.AgostoDolaresP = 0;
    this.SeptiembreDolaresP = 0;
    this.OctubreDolaresP = 0;
    this.NoviembreDolaresP = 0;
    this.DiciembreDolaresP = 0;
    this.EneroSolesG = 0;
    this.FebreroSolesG = 0;
    this.MarzoSolesG = 0;
    this.AbrilSolesG = 0;
    this.MayoSolesG = 0;
    this.JunioSolesG = 0;
    this.JulioSolesG = 0;
    this.AgostoSolesG = 0;
    this.SeptiembreSolesG = 0;
    this.OctubreSolesG = 0;
    this.NoviembreSolesG = 0;
    this.DiciembreSolesG = 0;
    this.EneroDolaresG = 0;
    this.FebreroDolaresG = 0;
    this.MarzoDolaresG = 0;
    this.AbrilDolaresG = 0;
    this.MayoDolaresG = 0;
    this.JunioDolaresG = 0;
    this.JulioDolaresG = 0;
    this.AgostoDolaresG = 0;
    this.SeptiembreDolaresG = 0;
    this.OctubreDolaresG = 0;
    this.NoviembreDolaresG = 0;
    this.DiciembreDolaresG = 0;
    this.Codigo_Proyecto = '';
    this.Ano = 0;
    this.Movimiento = false;
  }
}